Track EventBus subscriptions, send Inventory events
The primary reason for the EventBus rewrite was to make it provide a
subscription handle that can be used to cancel it later.
The rest of the changes just make it a bit easier to use the bus: for
instance, the listener's event doesn't need to be downcast anymore.

An Inventory pushes events to the bus to notify listeners of changes in
its contents.
ocelot-desktop uses that to keep track of updates not caused by the user
(such as when you disk_drive.eject() a floppy).